Implement framebuffer viewing using ImageView and update README.
The final release of MonoNX, finally has framebuffer viewing.

Updated README to show discontinuation.
