+++ README.md
@@ -1,9 +1,11 @@
-# MonoNX
+# NOTICE: This application has been discontinued (and support will no longer be given to this application, use at your own risk!), all development efforts have been moved to our new emulator [Skyline](https://github.com/skyline-emu/skyline).
+
+## MonoNX
 ![ILogo](Common/Images/MonoNXRound128.png)
 
 An experimental Nintendo Switch emulator for Android!
 
-# Usage
+## Usage
 
 You will need a device with the following, minimal specifications: 
 * A device running Android 5.1 or greater.
@@ -11,12 +13,12 @@ You will need a device with the following, minimal specifications:
 
 In order to play or use any form of Homebrew, the Application (as of now) will need you to configure Storage permissions manually; unless your copy of Android asks for permissions automatically.
 
-# Credits
+## Credits
 * Ryujinx Team - For the base of MonoNX, without them this would not be possible! Show them some love over at their [GitHub](https://github.com/Ryujinx/Ryujinx).
 * Xamarin - For allowing us C# developers to program on Android.
 * Dr.Hacknik - Logo design.
 
-# Support 
+## Support 
 You may join us on our Discord server, [here](https://discord.gg/XnbXNQM). You can get help, ask questions or help contribute there.
 
 The official GBAtemp thread can be found, [here](https://gbatemp.net/threads/mononx-an-experimental-nintendo-switch-emulator-for-android.531427/)
